requirements.txt
tabula-py
JPype1
pandas
streamlit